refactor: routers folder - move manual_fixes into utility folder
